break TIndexBuilder subclasses into separate files
